requirements.txt
langchain
python-dotenv
langchain-openai
notebook
streamlit
pandas
langchain_groq
langchain-google-genai
pypdf
faiss-cpu
langchain_community
structlog
PyMuPDF
-e .
